Degrees of Deathly Indifference
Matching Dr. L. Pedersen with any particular service need is pretty much an ideal-prize standard of her shield-ethic code, and this is why she is hard to place outside of a Doctoral degree in law and health. Her services touch all the colors of the rainbow, and her intensive times spent training with NASA as her guidestar through application simulations on all vast arrays of apps have built her up into a heavy weighted champ of stature matching close to the president of the U.S. 

Mr. Obama is a rolemodel of Mrs. Pedersen, and his status is non-comparible without her included along side her efforts which actually fueled his sudden change of status in "Obamacare" mutating it into the "healthcare reformation act" on behalf of inspirational letters, or e-mails, Mrs. Pedersen had sent in deadly precise outreach to Mr. Obama Jr. himself at the WhiteHouse page starting back at the beginning of this year/end of last year. Her cries and pleas to the president strike-forced a deep fire of fuel for his campaign cause, inviting him to finish out his presidential time in office along side her as a shepherd for good-will and change. This is the change you see today taking place, Mrs. Pedersen fueled and foughts for the suffering of our U.S. Veterans, and this sets her farther apart from any doctor or lawyer in history.